manifest: Remove to use specific system smack label
In general, most shell script has '_' floor label of floor domain on
Three-domain SMACK of Tizen 3. BTW, this script seems to be used by
something process included in system domain like dump app on tizen 2.4
but i'm not sure about that.

Now this script is not used by anyone on Tizen 3, so just remove to use
specific system smack label in manifest of this package by guide of
security team.

Change-Id: Ice80cd815ca0872aa50ec15ee2aa85bae3db4a7e
Signed-off-by: Joonyoung Shim <jy0922.shim@samsung.com>